Starting Static Analysis
Added Larastan to the package and analyzing the code.
Going with level 5 to start with.
Adding type information.
Casting some variable to match the return type.
Added a baseline for errors not fixing at this point. Mostly calls to methods that come from traits.
Left 2 errors that I want to get fixed.