Larastan Level 8
Working through the errors as I increase the phpstan level up to 8.

Narrowing some Types to identify what is being put into collect().

searchConfig should return a string and not string|null. If the not defined then 'simple' will be the text search configuration.

'plainquery' is the defaultQueryMethod. Changed the default from 'plain' to 'plainquery'.